more of the same with get_objects
