+++ rets/session.py
@@ -5,8 +5,8 @@ import logging
 import sys
 from rets.exceptions import InvalidSearch, RETSException, NotLoggedIn
 from rets.utils.get_object import GetObject
-from rets.parsers.get_object.multiple import MultipleObjectParser
-from rets.parsers.get_object.single import SingleObjectParser
+from rets.parsers.get_object import MultipleObjectParser
+from rets.parsers.get_object import SingleObjectParser
 from rets.parsers.search import OneXSearchCursor
 from rets.parsers.login import OneXLogin
 from rets.parsers.metadata import Metadata
